Use sse request to fetch project data
