+++ portal/src/routes/(authenticated)/projects/[id=idNumber]/+page.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import { page } from '$app/state';
   import IconContainer from '$lib/components/IconContainer.svelte';
   import Tooltip from '$lib/components/Tooltip.svelte';
   import { l10nMap, tryLocalizeName } from '$lib/locales.svelte';
@@ -7,180 +8,195 @@
   import ProjectActionMenu from '$lib/projects/components/ProjectActionMenu.svelte';
   import { byName } from '$lib/utils/sorting';
   import { getRelativeTime } from '$lib/utils/time';
-  import type { PageData } from './$types';
+  import type { Readable } from 'svelte/store';
+  import { source } from 'sveltekit-sse';
   import { Authors, OwnerGroup, Reviewers, Settings } from './forms';
   import { AddProduct } from './modals';
   import ProductCard from './ProductCard.svelte';
+  import type { ProjectDataSSE } from './sse/+server';
 
-  interface Props {
-    data: PageData;
-  }
-
-  let { data }: Props = $props();
+  const { data } = $props();
 
   let addProductModal: HTMLDialogElement | undefined = $state(undefined);
+
+  const projectData: Readable<ProjectDataSSE> = source(`${page.params.id}/sse`)
+    .select('projectData')
+    .json();
 </script>
 
 <div class="w-full max-w-6xl mx-auto relative">
-  <!-- Header -->
-  <div class="flex p-6">
-    <div class="shrink">
-      <h1 class="p-0">
-        {data.project?.Name}
-      </h1>
-      <div>
-        <span class="font-bold">
-          {data.project?.IsPublic ? m.project_public() : m.project_private()}
-        </span>
-        <span>-</span>
-        <span>
-          {m.project_createdOn()}
-          <Tooltip tip={data.project?.DateCreated?.toLocaleString(getLocale())}>
-            {data.project?.DateCreated ? getRelativeTime(data.project?.DateCreated) : 'null'}
-          </Tooltip>
-        </span>
-      </div>
-      {#if data.project?.DateArchived}
-        <span>
-          {m.project_archivedOn()}
-          <Tooltip tip={data.project?.DateArchived?.toLocaleString(getLocale())}>
-            {data.project?.DateArchived ? getRelativeTime(data.project?.DateArchived) : 'null'}
-          </Tooltip>
-        </span>
-      {/if}
+  {#if !$projectData}
+    <div class="flex justify-center items-center h-64">
+      <span class="loading loading-spinner loading-lg"></span>
     </div>
-    <div class="grow">
-      <Tooltip className="tooltip-bottom" tip={m.project_editProject()}>
-        <a
-          href={localizeHref(`/projects/${data.project?.Id}/edit`)}
-          title={m.project_editProject()}
-        >
-          <IconContainer width="24" icon="mdi:pencil" />
-        </a>
-      </Tooltip>
-    </div>
-    <div class="shrink">
-      <ProjectActionMenu
-        data={data.actionForm}
-        project={data.project}
-        userGroups={data.userGroups}
-        orgId={data.project.Organization.Id}
-      />
-    </div>
-  </div>
-  <div class="grid maingrid w-full p-4 pb-0">
-    <div class="mainarea min-w-0">
-      <!-- Details -->
-      <h2 class="pl-0">{m.project_details_title()}</h2>
-      <div>
-        <div class="gridcont grid gap-x-6 gap-y-2">
-          <div class="flex place-content-between">
-            <span>
-              <IconContainer icon="ph:globe" width={20} />
-              {m.project_details_language()}:
-            </span>
-            <span>
-              {data.project?.Language} ({tryLocalizeName(
-                data.langtags,
-                l10nMap.value,
-                getLocale(),
-                data.project.Language ?? ''
-              )})
-            </span>
-          </div>
-          <div class="flex place-content-between">
-            <span>{m.project_details_type()}:</span>
-            <span>{data.project?.ApplicationType.Description}</span>
-          </div>
+  {:else}
+    <!-- Header -->
+    <div class="flex p-6">
+      <div class="shrink">
+        <h1 class="p-0">
+          {$projectData?.project?.Name}
+        </h1>
+        <div>
+          <span class="font-bold">
+            {$projectData?.project?.IsPublic ? m.project_public() : m.project_private()}
+          </span>
+          <span>-</span>
+          <span>
+            {m.project_createdOn()}
+            <Tooltip tip={$projectData?.project?.DateCreated?.toLocaleString(getLocale())}>
+              {$projectData?.project?.DateCreated
+                ? getRelativeTime($projectData?.project?.DateCreated)
+                : 'null'}
+            </Tooltip>
+          </span>
         </div>
-        <div class="my-4">
-          <span>{m.project_description()}:</span>
-          <br />
-          <p>{data.project?.Description}</p>
-        </div>
-        {#if data.project?.WorkflowProjectUrl}
-          <div>
-            <span>{m.project_location()}:</span>
-            <br />
-            <p class="rounded-md text-nowrap overflow-x-scroll bg-base-200 p-3 pt-2 mt-2">
-              {data.project?.WorkflowProjectUrl}
-            </p>
-          </div>
+        {#if $projectData?.project?.DateArchived}
+          <span>
+            {m.project_archivedOn()}
+            <Tooltip tip={$projectData?.project?.DateArchived?.toLocaleString(getLocale())}>
+              {$projectData?.project?.DateArchived
+                ? getRelativeTime($projectData?.project?.DateArchived)
+                : 'null'}
+            </Tooltip>
+          </span>
         {/if}
       </div>
-      <!-- Product List Header -->
-      <div class="flex flex-row place-content-between items-end">
-        <div>
-          <h2 class="pl-0">{m.products_title()}</h2>
-          <div class="mb-2">
-            <span class="italic">{m.products_definition()}</span>
-          </div>
-        </div>
-        <button
-          class="btn btn-outline"
-          onclick={() => addProductModal?.showModal()}
-          disabled={!(data.productsToAdd.length && data.project.WorkflowProjectUrl)}
-        >
-          {m.products_add()}
-        </button>
-        <AddProduct
-          bind:modal={addProductModal}
-          prodDefs={data.productsToAdd}
-          stores={data.stores}
-          endpoint="addProduct"
+      <div class="grow">
+        <Tooltip className="tooltip-bottom" tip={m.project_editProject()}>
+          <a
+            href={localizeHref(`/projects/${$projectData?.project?.Id}/edit`)}
+            title={m.project_editProject()}
+          >
+            <IconContainer width="24" icon="mdi:pencil" />
+          </a>
+        </Tooltip>
+      </div>
+      <div class="shrink">
+        <ProjectActionMenu
+          data={data.actionForm}
+          project={$projectData?.project}
+          userGroups={$projectData?.userGroups}
+          orgId={$projectData?.project.Organization.Id}
         />
       </div>
-      <!-- Products List -->
-      <div>
-        {#if !data.project?.Products.length}
-          {m.projectTable_noProducts()}
-        {:else}
-          {#each data.project.Products.toSorted( (a, b) => byName(a.ProductDefinition, b.ProductDefinition, getLocale()) ) as product}
-            <ProductCard
-              {product}
-              project={data.project}
-              actionEndpoint="productAction"
-              deleteEndpoint="deleteProduct"
-              updateEndpoint="updateProduct"
-            />
-          {/each}
-        {/if}
-        <div class="divider"></div>
+    </div>
+    <div class="grid maingrid w-full p-4 pb-0">
+      <div class="mainarea min-w-0">
+        <!-- Details -->
+        <h2 class="pl-0">{m.project_details_title()}</h2>
+        <div>
+          <div class="gridcont grid gap-x-6 gap-y-2">
+            <div class="flex place-content-between">
+              <span>
+                <IconContainer icon="ph:globe" width={20} />
+                {m.project_details_language()}:
+              </span>
+              <span>
+                {$projectData?.project?.Language} ({tryLocalizeName(
+                  data.langtags,
+                  l10nMap.value,
+                  getLocale(),
+                  $projectData?.project.Language ?? ''
+                )})
+              </span>
+            </div>
+            <div class="flex place-content-between">
+              <span>{m.project_details_type()}:</span>
+              <span>{$projectData?.project?.ApplicationType.Description}</span>
+            </div>
+          </div>
+          <div class="my-4">
+            <span>{m.project_description()}:</span>
+            <br />
+            <p>{$projectData?.project?.Description}</p>
+          </div>
+          {#if $projectData?.project?.WorkflowProjectUrl}
+            <div>
+              <span>{m.project_location()}:</span>
+              <br />
+              <p class="rounded-md text-nowrap overflow-x-scroll bg-base-200 p-3 pt-2 mt-2">
+                {$projectData?.project?.WorkflowProjectUrl}
+              </p>
+            </div>
+          {/if}
+        </div>
+        <!-- Product List Header -->
+        <div class="flex flex-row place-content-between items-end">
+          <div>
+            <h2 class="pl-0">{m.products_title()}</h2>
+            <div class="mb-2">
+              <span class="italic">{m.products_definition()}</span>
+            </div>
+          </div>
+          <button
+            class="btn btn-outline"
+            onclick={() => addProductModal?.showModal()}
+            disabled={!(
+              $projectData?.productsToAdd.length && $projectData?.project.WorkflowProjectUrl
+            )}
+          >
+            {m.products_add()}
+          </button>
+          <AddProduct
+            bind:modal={addProductModal}
+            prodDefs={$projectData?.productsToAdd}
+            stores={$projectData?.stores}
+            endpoint="addProduct"
+          />
+        </div>
+        <!-- Products List -->
+        <div>
+          {#if !$projectData?.project?.Products.length}
+            {m.projectTable_noProducts()}
+          {:else}
+            {#each $projectData?.project.Products.toSorted( (a, b) => byName(a.ProductDefinition, b.ProductDefinition, getLocale()) ) as product}
+              <ProductCard
+                {product}
+                project={$projectData?.project}
+                actionEndpoint="productAction"
+                deleteEndpoint="deleteProduct"
+                updateEndpoint="updateProduct"
+              />
+            {/each}
+          {/if}
+          <div class="divider"></div>
+        </div>
+      </div>
+      <!-- Settings -->
+      <div class="settingsarea my-4">
+        <Settings
+          project={$projectData?.project}
+          publicEndpoint="toggleVisibility"
+          downloadEndpoint="toggleDownload"
+        />
+      </div>
+      <!-- Sidebar Settings -->
+      <div class="space-y-2 min-w-0 flex-auto sidebararea">
+        <OwnerGroup
+          project={$projectData?.project}
+          users={$projectData?.possibleProjectOwners}
+          groups={$projectData?.possibleGroups}
+          orgName={data.organizations.find((o) => o.Id === $projectData?.project.Organization.Id)
+            ?.Name}
+          endpoint="editOwnerGroup"
+        />
+        <Authors
+          group={$projectData?.project.Group}
+          projectAuthors={$projectData?.project.Authors}
+          availableAuthors={$projectData?.authorsToAdd}
+          formData={data.authorForm}
+          createEndpoint="addAuthor"
+          deleteEndpoint="deleteAuthor"
+        />
+        <Reviewers
+          reviewers={$projectData?.project.Reviewers}
+          formData={data.reviewerForm}
+          createEndpoint="addReviewer"
+          deleteEndpoint="deleteReviewer"
+        />
       </div>
     </div>
-    <!-- Settings -->
-    <div class="settingsarea my-4">
-      <Settings
-        project={data.project}
-        publicEndpoint="toggleVisibility"
-        downloadEndpoint="toggleDownload"
-      />
-    </div>
-    <!-- Sidebar Settings -->
-    <div class="space-y-2 min-w-0 flex-auto sidebararea">
-      <OwnerGroup
-        project={data.project}
-        users={data.possibleProjectOwners}
-        groups={data.possibleGroups}
-        orgName={data.organizations.find((o) => o.Id === data.project.Organization.Id)?.Name}
-        endpoint="editOwnerGroup"
-      />
-      <Authors
-        group={data.project.Group}
-        projectAuthors={data.project.Authors}
-        availableAuthors={data.authorsToAdd}
-        formData={data.authorForm}
-        createEndpoint="addAuthor"
-        deleteEndpoint="deleteAuthor"
-      />
-      <Reviewers
-        reviewers={data.project.Reviewers}
-        formData={data.reviewerForm}
-        createEndpoint="addReviewer"
-        deleteEndpoint="deleteReviewer"
-      />
-    </div>
-  </div>
+  {/if}
 </div>
 
 <style>
